fix: stepper state관리 context api로 변경
