--- src/components/stepper/store/atom.ts
@@ -1,11 +1,0 @@
-import { atom } from 'jotai';
-
-type StepperAtomType = {
-  activeStep: number;
-  totalCount: number;
-};
-
-export const stepperAtom = atom<StepperAtomType>({
-  activeStep: 0,
-  totalCount: 0,
-});
